+++ CHANGELOG.md
@@ -20,6 +20,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Button*` font sizes updated to match design specifications
 - `theme.colors` shape has been significantly refactored (documentation updates to follow)
+- `Banner` is now `MessageBar`
+  - Major style changes to be more Material-esque
+  - `warning` is now `warn`
+  - `error` is now `critical`
+  - `info` is now `inform`
+  - `confirmation` is now `positive`
+- ``
+- `Badge` - `default` intent is now `key`
 - `CheckboxGroup` and `RadioGroup` `name` is now optional
 - Updated `LookerLogo` and `LogoRings` to match new branding